backend/requirements.txt
# Flask 核心
Flask==3.0.0
Werkzeug==3.0.1

# 数据库
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
SQLAlchemy>=2.0.36

# JWT 认证
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0

# MQTT 客户端
paho-mqtt==1.6.1

# 密码加密
bcrypt==4.1.2

# 跨域支持
Flask-CORS==4.0.0

# 工具库
python-dotenv==1.0.0
